Mark the expression to avoid reprocessing the same expression.
In some circumstances, Babel will retraverse the same path. This is
generally caused by other plugins needing to process the path also
(which our tests do not currently excercise) which then causes a
second traversal (to ensure the our plugin has access to the changed
expression).

When this retraversal happens, we end up pushing another entry into
`Builder#expressions` with the same `path` which _roughly_ equates to
the "last one wins".  Prior to recent changes to avoid duplicating the
statements predicate this was fine, but after that change we now mutate
the `expressions.arguments` to replace the predicate with `false`.
This means that the second traversal path takes that `false` literal
and uses it as the predicate itself (causing the "real" predicate to be
completely lost).

This change marks a given expression as "seen" so that we do not
reprocess an already processed expression.