Ensure import is removed when other tooling uses the same module.
In the default configuration of ember-cli-babel both `babel-plugin-debug-macros`
and `babel-plugin-ember-modules-api-polyfill` handle specific specifiers out of
`@ember/debug`. Prior to this change this package would not properly remove the
import declaration (even though it had no specifiers left after both plugins had
ran), because the `isImportRemovable` flag is eagerly cached.

After this change, the import is removed if the only remaining specifiers are
the ones that debug-macros is removing.