Add a mode that converts to @embroider/macros
This is intended as a compatibility path for packages (like ember-source) that make heavy use of babel-plugin-debug-macros and want to become valid V2 addons.

A v2 addon can't bring its own custom build pipeline along into apps. If they want to make compile-time decisions, they need to do it through the standardized `@embroider/macros`.

This implementation does the smallest, simplest thing, which is replacing the hard-coded boolean we used to emit with a call to `isDevelopingApp()` from '@embroider/macros`. It relies on the same kind of dead-code elimination that this babel plugin has always relied on.